[FIX] Alternative Dart implementation
Hi,

Great job for the plugin! Unfortunately because of conflict between pdfx and your plugin with Pigeon I could not go for your plugin so I had to rely on the dart implementation. Base64 would not give me the same md5sum so its better to just use .toString() method.